requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr<2.0,>=0.11
python-dateutil>=2.4.2
cliff>=1.13.0  # Apache-2.0
python-heatclient>=0.6.0
python-openstackclient>=1.5.0
# grandalf
svgwrite
